Change TileMarkerType editing to a panel for clarity.
Removed redundant and test code.

Renamed some classes and variables to more appropriate names.

Changed the editing section of a TileMarkerType from a dropdown in the UIList to its own panel as the scrolling was confusing (the expansion counted as an item in a list so a single scroll would scroll the whole expanded item, just as it would a non-expanded item so having a mix of these in the list caused different scroll rates making it confusing to follow).

Added missing properties to TileMarkerType.