Add material filtering to preferences, assign materials on gpencil creation and rename stroke_material to gpencil_material.
Added polling to TileMarkerType to filter for the allowed materials (gpencil or not).

Renamed stroke_material to gpencil_material in most places to be consistent with the Blender-assigned name.

Made it so materials are checked and assigned when the gpencil is created so that it will have the up to date preferences.