Add tile_size to config and update config UI layout.
The UI now uses sub panels to keep it all together and make it clearer it all belongs together.

The tile size is separate from everything else as it is not a material, unlike them.

The erase material was moved above the marker types and separated a bit to clearly show the difference.

Added tooltips to marker type materials.